feat(rxjsracecondition): fixed race condition
Observable is now subscribed to within the `TopicModalComponent`
template instead of the AppComponent. This means the fetch only occurs
on opening of the modal.

I decided to go with passing the observable through the
`MAT_DIALOG_DATA`. An alternative would have been to just use the
`TopicService` within the `TopicDialogComponent`. However that would
involve a lot more code to allow it still be decoupled from the source.
Although it's probably more flexible down the line.