feat(decoupling): add solution for challenge 33
create an injection token representing a signal
for the button status.

In this solution the injection token is the signal
itself. If it would be a service, the token wouldn't
be necessary.

Since the directive needs access to a writable
signal, it needs to come up with an additional
hidden service, that contains the actual signal.

The injection token gets the signal from that
service and thus we can achieve a readonly signal.