refactor: improve card component for reusability
In the current state, the application has a functional dashboard with teacher and student cards. However, the developer experience is suboptimal due to the need for modifications in card.component.ts every time a new card is implemented. This is not scalable, especially for shared components across multiple projects.

To address this, I've refactored the code to introduce a CardComponent that can be customized without altering the core component. This improvement enhances reusability and maintainability. With this refactor in place, I'm now ready to proceed with implementing the CityCardComponent without any need to touch the CardComponent.